# 🎯 PROJECT OPTIMIZATION 2025: 47% File Reduction + 300% Faster Setup